azure-function-project/requirements.txt
azure-functions
openai
tqdm
httpx
azure-cosmos